% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage_functions.R
\name{df_S_lasso}
\alias{df_S_lasso}
\title{Derivative of soft thresholding a core array.}
\usage{
df_S_lasso(S, lambda)
}
\arguments{
\item{S}{An array of numerics.}

\item{lambda}{A numeric. The threshholding parameter.}
}
\description{
Derivative of soft thresholding a core array.
}
\author{
David Gerard.
}
\seealso{
\code{\link{df_S_lasso}}.
}

