% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage_functions.R
\name{df_S_lasso_mult}
\alias{df_S_lasso_mult}
\title{Derivative of scaling and soft thresholding a core array.}
\usage{
df_S_lasso_mult(S, params)
}
\arguments{
\item{S}{An array of numerics.}

\item{params}{A vector of length 2 of numerics. \code{param[1]} is the
scaling parameter, \code{param[2]} is the threshholding parameter.}
}
\description{
Derivative of scaling and soft thresholding a core array.
}
\author{
David Gerard.
}
\seealso{
\code{\link{df_S_lasso_mult}}.
}

