% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_sure.R
\name{df_truncate}
\alias{df_truncate}
\title{Derivative of truncation spectral function.}
\usage{
df_truncate(x, r)
}
\arguments{
\item{x}{A vector of numerics.}

\item{r}{A positive integer.}
}
\description{
Derivative of truncation spectral function.
}
\author{
David Gerard.
}
\seealso{
\code{\link{f_truncate}}.
}

