% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_est.R
\name{efron_morris}
\alias{efron_morris}
\title{Efron-Morris estimator.}
\usage{
efron_morris(X, mode_index = 1, tau2 = 1)
}
\arguments{
\item{X}{An array of numerics. The data tensor.}

\item{mode_index}{The mode upon which to apply the Efron-Morris estimator.
Defaults to the first mode.}

\item{tau2}{A positive numeric. The variance. Assumed known and defaults to
1.}
}
\description{
Efron-Morris estimator.
}
\author{
David Gerard.
}
\references{
Efron, B., & Morris, C. (1972).
  \href{http://biomet.oxfordjournals.org/content/59/2/335.short}{Empirical
  Bayes on vector observations: An extension of Stein's method}.
  \emph{Biometrika}, 59(2), 335-347.
}

