% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage_functions.R
\name{f_lasso}
\alias{f_lasso}
\title{Soft thresholding shrinkage function. Same as lasso for spectral shrinkage.}
\usage{
f_lasso(x, lambda)
}
\arguments{
\item{x}{A vector of numerics.}

\item{lambda}{A numeric. The threshholding parameter.}
}
\description{
Soft thresholding shrinkage function. Same as lasso for spectral shrinkage.
}
\seealso{
\code{\link{df_lasso}}.
}
\author{
David Gerard.
}
