% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage_functions.R
\name{f_truncate}
\alias{f_truncate}
\title{Truncation spectral function.}
\usage{
f_truncate(x, r)
}
\arguments{
\item{x}{A vector of numerics.}

\item{r}{A positive integer.}
}
\description{
Truncation spectral function.
}
\author{
David Gerard.
}
\seealso{
\code{\link{df_truncate}}.
}

