% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_sure.R
\name{get_basics}
\alias{get_basics}
\title{Calculates necessary components to calculate the SURE for estimators that
shrink the individual elements of the core array.}
\usage{
get_basics(X)
}
\arguments{
\item{X}{An array of numerics. The tensor data.}
}
\value{
A list of objects meant to be fed into
  \code{\link{sure_given_basics}} to calculate the SURE for higher-order
  spectral estimators that shrink the core array directly.
}
\description{
Calculates necessary components to calculate the SURE for estimators that
shrink the individual elements of the core array.
}
\author{
David Gerard.
}
\references{
Gerard, D., & Hoff, P. (2015).
  \href{http://arxiv.org/abs/1505.02114}{Adaptive Higher-order Spectral
  Estimators}. \emph{arXiv preprint} arXiv:1505.02114.
}
\seealso{
\code{\link{sure_given_basics}}, \code{\link{sure_S}}
}

