% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_specific_sure.R
\name{get_c}
\alias{get_c}
\title{Outputs the 'C' array from Gerard and Hoff (2015), along with the HOSVD of
the data tensor.}
\usage{
get_c(X)
}
\arguments{
\item{X}{An array of numerics. The data.}
}
\value{
\code{C_array} An array of numerics. The "C" array from Gerard and
  Hoff (2015).

  \code{hosvd_x} A list containing the higher-order SVD of \code{X}.
}
\description{
This is necessary to calculate the SURE for higher-order spectral
estimators.
}
\details{
this function only calculates C and the HOSVD so that calculating SURE is
faster when doing it over and over again
}
\references{
Gerard, D., & Hoff, P. (2015).
  \href{http://arxiv.org/abs/1505.02114}{Adaptive Higher-order Spectral
  Estimators}. \emph{arXiv preprint} arXiv:1505.02114.
}
\seealso{
\code{\link{diverge_given_c}} for calculating the divergence of
  higher-order spectral estimators using the output of \code{get_c}.

  \code{\link{sure_given_c}} for calculating the SURE given the output of
  \code{get_c}.

  \code{\link{sure}} for a wrapper for \code{get_c} and
  \code{\link{sure_given_c}}.

  \code{\link{soft_coord}} for a coordinate descent algorithm for finding
  the optimal sure using the output from \code{get_c}.
}
\author{
David Gerard.
}
