% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cichocki_meet.R
\name{get_trunc_est}
\alias{get_trunc_est}
\title{Return the estimate from the truncated HOSVD given the hosvd and the rank.}
\usage{
get_trunc_est(fit_hosvd, rank_vec, return_est = TRUE)
}
\arguments{
\item{fit_hosvd}{The output from \code{\link{hosvd_full}}}

\item{rank_vec}{The multilinear rank.}

\item{return_est}{A logical. Should we return the estimate (\code{TRUE}),
or not (\code{FALSE})?}
}
\description{
Return the estimate from the truncated HOSVD given the hosvd and the rank.
}
\author{
David Gerard
}
