% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hosvd_full}
\alias{hosvd_full}
\title{Higher-order SVD using same signed eigenvectors as matrix SVD's.}
\usage{
hosvd_full(Y)
}
\arguments{
\item{Y}{An array of numerics.}
}
\value{
\code{U} A list of matrices with orthonormal columns. Each matrix
  contains the mode-specific singular vectors of its mode.

  \code{D} A list of vectors of numerics. The \eqn{k}th vector contains the mode-specific singular
  values of the \eqn{k}th matricization of \code{Y}

  \code{S} An all-orthogonal array. This is the core array from the HOSVD.
}
\description{
Calculates the left singular vectors of each matrix unfolding of an array,
then calculates the core array. The resulting output is a Tucker
decomposition.
}
\details{
\code{Y} is equal to \code{atrans(S, U)}, up to numerical accuracy.

This function differs
from the \code{hosvd} function in the package \code{tensr} only in  (1) the sign
conditions on the core array and (2) it will also return the mode-specific singular values.

\href{http://epubs.siam.org/doi/abs/10.1137/S0895479896305696}{ De Lathauwer
et. al. (2000)}.
}
\references{
De Lathauwer, L., De Moor, B., & Vandewalle, J. (2000).
  \href{http://epubs.siam.org/doi/abs/10.1137/S0895479896305696}{A
  multilinear singular value decomposition}. \emph{SIAM journal on Matrix
  Analysis and Applications}, 21(4), 1253-1278.
}
\author{
David Gerard.
}
