% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_sure.R
\name{listSum}
\alias{listSum}
\title{Sums elements in a list.}
\usage{
listSum(D)
}
\arguments{
\item{D}{A list of summable elements.}
}
\value{
\code{list_sum} The sum of the elements in \code{D}.
}
\description{
Sums elements in a list.
}
\author{
David Gerard.
}

