% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cichocki_meet.R
\name{mdl_eigen}
\alias{mdl_eigen}
\title{Returns the minimum description length given the eigenvalues.}
\usage{
mdl_eigen(lambda_vec, rho, I_nbar)
}
\arguments{
\item{lambda_vec}{A numeric vector of eigenvalues.}

\item{rho}{The estimated rho parameter from Yokota, Lee, and Cichocki (2017)}

\item{I_nbar}{The product of the other dimensions.}
}
\value{
A list with two elements: \code{rank_est} the estimated rank; and \code{dl} a data-frame
    with the description lengths for each rank from \code{0} to \code{length(lambda_vec) - 1}.
}
\description{
Returns the minimum description length given the eigenvalues.
}
\references{
Yokota, Tatsuya, Namgil Lee, and Andrzej Cichocki. "Robust multilinear tensor rank estimation using higher order singular value decomposition and information criteria." IEEE Transactions on Signal Processing 65.5 (2017): 1196-1206. DOI: 10.1109/TSP.2016.2620965
}
\seealso{
\code{\link{meet}}
}
\author{
David Gerard
}
