% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cichocki_meet.R
\name{meet}
\alias{meet}
\title{Implements the Modified eivgenvalues estimator for Tucker rank determination (MEET) algorithm
from Yokota, Lee, and Cichocki (2017).}
\usage{
meet(Y, tau2 = NULL)
}
\arguments{
\item{Y}{A numeric array.}

\item{tau2}{The known variance. Will be estimated if missing. Not supported yet.}
}
\value{
A list with two elements: \code{rank} the estimated multilinear rank of the tensor and
    \code{sigma} a vector of the estimated variance when using each mode.
}
\description{
The model is \deqn{Y = A + E,} where A is a low-rank matrix and E is
the noise matrix is i.i.d. N(0,s) entries. s is not assumed to be known.
}
\references{
Yokota, Tatsuya, Namgil Lee, and Andrzej Cichocki. "Robust multilinear tensor rank estimation using higher order singular value decomposition and information criteria." IEEE Transactions on Signal Processing 65.5 (2017): 1196-1206. DOI: 10.1109/TSP.2016.2620965
}
\author{
David Gerard
}
