% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_sure.R
\name{min_ef}
\alias{min_ef}
\title{Min Efron-Morris estimator.}
\usage{
min_ef(X, tau2 = 1)
}
\arguments{
\item{X}{An array of numerics. The data tensor.}

\item{tau2}{A positive numeric. The variance. Assumed known and defaults to
1.}
}
\description{
Min Efron-Morris estimator.
}
\author{
David Gerard
}
\references{
Efron, B., & Morris, C. (1972).
  \href{http://biomet.oxfordjournals.org/content/59/2/335.short}{Empirical
  Bayes on vector observations: An extension of Stein's method}.
  \emph{Biometrika}, 59(2), 335-347.
}

