% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cichocki_meet.R
\name{mode_mdl}
\alias{mode_mdl}
\title{Mode-wise MDL criterion.}
\usage{
mode_mdl(Y, return_est = TRUE)
}
\arguments{
\item{Y}{A numeric array.}

\item{return_est}{A logical. Should we return the estimate of the mean?}
}
\description{
Mode-wise MDL criterion.
}
\author{
David Gerard
}
