% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_sure.R
\name{pad_circ}
\alias{pad_circ}
\title{Adds a circular padding after matrix.}
\usage{
pad_circ(X, padsize)
}
\arguments{
\item{X}{A matrix of numerics.}

\item{padsize}{A positive integer. The number of columns/rows to add to
\code{X}.}
}
\description{
Adds a circular padding after matrix.
}
\author{
David Gerard.
}

