% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cichocki_meet.R
\name{par_analysis_wrapper}
\alias{par_analysis_wrapper}
\title{A wrapper for \code{\link[sva]{num.sv}} for each mode.}
\usage{
par_analysis_wrapper(Y, return_est = TRUE)
}
\arguments{
\item{Y}{An array of numerics.}

\item{return_est}{A logical. Should we return the truncated HOSVD
(\code{TRUE}) or not (\code{FALSE})?}
}
\description{
This function requires \code{sva}. To install \code{sva}, run in R:
\code{source("https://bioconductor.org/biocLite.R")} then
\code{biocLite("sva")}
}
\author{
David Gerard
}
