% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pos_part}
\alias{pos_part}
\title{Positive part function.}
\usage{
pos_part(x)
}
\arguments{
\item{x}{A vector of numerics.}
}
\description{
Returns a vector whose elements are the positive parts of the elements of
\code{x}.
}
\author{
David Gerard.
}

