% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pos_part_2}
\alias{pos_part_2}
\title{Positive part function.}
\usage{
pos_part_2(X)
}
\arguments{
\item{X}{A vector, matrix, or array of numerics.}
}
\description{
Returns an object whose elements are the positive parts of the elements of
\code{X}.
}
\author{
David Gerard.
}

