% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cichocki_meet.R
\name{score_ylc}
\alias{score_ylc}
\title{SCORE method from Yokota, Lee, and Cichoki (2017).}
\usage{
score_ylc(Y, rho = 0.01, sampmin = min(c(dim(Y), 5)), return_est = TRUE,
  verbose = FALSE)
}
\arguments{
\item{Y}{An array of numerics.}

\item{rho}{The proportion of the core array to use to calcuate the
mode-specific modified singular values.}

\item{sampmin}{The minimum number of columns to of the core to use
in calculating the mode-specific modified singular values.}

\item{return_est}{A logical. Should we return the truncated HOSVD
(\code{TRUE}) or not (\code{FALSE})?}

\item{verbose}{A logical. Should we print a lot?}
}
\description{
Implements the "Sparse CORE" method from Yokota, Lee, and Cichocki (2017).
The idea is that one uses only a subset of the core array from the HOSVD
to calculate the "modified" singular values. Then one uses the minimum
description length (see \code{\link{mdl_eigen}}) criteria to choose the rank
of the array. The porportion of the core should be small. We have one percent
as the default as suggested in their paper, but my guess is that this should
actually depend on the rank of the mean tensor.
}
\references{
Yokota, Tatsuya, Namgil Lee, and Andrzej Cichocki. "Robust multilinear tensor rank estimation using higher order singular value decomposition and information criteria." IEEE Transactions on Signal Processing 65.5 (2017): 1196-1206. DOI: 10.1109/TSP.2016.2620965
}
\author{
David Gerard
}
