% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_specific_sure.R
\name{sgd_given_c}
\alias{sgd_given_c}
\title{Runs a stochastic gradient descent to minimize SURE for higher-order
spectral estimators.}
\usage{
sgd_given_c(c_obj, sgd_lambda, sgd_lambda_c, c_init = 1, lambda_init = NULL,
  sgd_c = 1/2 + 0.001, itermax = 10000, tau2 = 1, print_current = TRUE,
  every_iter = 1000, alpha = 0.2, calc_final = TRUE)
}
\arguments{
\item{c_obj}{= Output from \code{\link{get_c}}.}

\item{sgd_lambda}{Lambda value from sgd for lambda from hose.}

\item{sgd_lambda_c}{Lambda value from sgd for c from hose.}

\item{c_init}{Initalization for c from hose.}

\item{lambda_init}{Initialization for lambda from hose.}

\item{sgd_c}{c value from sgd.}

\item{itermax}{Maximum number of iterations for sgd.}

\item{tau2}{Known variance.}

\item{print_current}{Should we print the results at each iteration?}

\item{every_iter}{If \code{print_current = TRUE}, then every
\code{every_iter} iteration will be printed.}

\item{alpha}{Burnin for final estimates.}

\item{calc_final}{Should we calculate the final sure and estimates?}
}
\description{
This function is in beta. Not sure if it works properly or is even
worthwhile considering that \code{\link{soft_coord}} works pretty well.
Also, we have to call \code{\link{get_c}} anyway, so not sure if this function
actually reduces the computational complexity.
}
\author{
David Gerard.
}
\references{
Gerard, D., & Hoff, P. (2015).
  \href{http://arxiv.org/abs/1505.02114}{Adaptive Higher-order Spectral
  Estimators}. \emph{arXiv preprint} arXiv:1505.02114.
}

