% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_ests.R
\name{sig_mp}
\alias{sig_mp}
\title{Estimate of the variance from Gavish and Donoho (2014) based on the
Marcenko-Pasture distribution.}
\usage{
sig_mp(dmed, N, p)
}
\arguments{
\item{dmed}{A positive numeric. The median singular value of the
data matrix}

\item{N}{A positive integer. The row dimension.}

\item{p}{A positive integer. The column dimension.}
}
\value{
\code{sig2_est} A positive numeric. The estimate of the
    variance of the data matrix.
}
\description{
Returns a normalized squared median singular value that is a
consistent estimate of the variance under a specific asymptotic
regime.
}
\details{
Under the asymptotic regime where the rank of the mean is fixed and
the signal to noise ratio is fixed, the squared median singular
value divided by the variance and the larger dimension size will
converge to the median of the Marcenko-Pasture distribution. Hence,
the squared median singular value divided by the larger dimension size
and the median of the Marcenko-Pastur distribution will converge to
the variance of the data matrix. This estimator works really well
when the rank is much less than the dimensions of the matrix and
really poorly otherwise.
}
\author{
David Gerard
}
\references{
Gavish, M., & Donoho,
    D. L. (2014). \href{http://ieeexplore.ieee.org/xpls/abs_all.jsp?arnumber=6846297&tag=1}{The
    optimal hard threshold for singular values is
    4/sqrt(3)}. Information Theory, IEEE Transactions on, 60(8),
    5040-5053.
}

