% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_ests.R
\name{sig_naive}
\alias{sig_naive}
\title{Simplest estimator for variance.}
\usage{
sig_naive(d, N, p, r)
}
\arguments{
\item{d}{A vector of positive numerics. The singular values of the
data matrix.}

\item{N}{A positive integer. The row dimension.}

\item{p}{A positive integer. The column dimension.}

\item{r}{A positive integer. The known rank of the matrix. Must be
less than \code{min(N, p)}}
}
\value{
\code{sig2_est} A positive numeric. The estimate of the
    variance.
}
\description{
An estimator for the variance of a matrix of independent random
variables when assuming the mean matrix is low rank with the rank
known.
}
\details{
Let \eqn{Y} be a matrix with row dimension \eqn{n} and column
dimension \eqn{p} where \eqn{n \ge p}. The model is \eqn{Y =
\Theta + \sigma E} where \eqn{\Theta} is low rank and \eqn{E}
contains independent elements with mean zero and variance one. The
MLE of \eqn{\sigma^2} is the sum of squares of the last \eqn{p - r}
singular values divided by \eqn{pn}. This estimator has negative
bias so Choi et al (2014) suggested to instead divide by \eqn{n(p -
r)}. This is the estimator implemented here. It seems to have
reasonable performance when the rank is small.  Though the biggest
drawback here is that you have to know the rank ahead of time.
}
\references{
Choi, Yunjin, Jonathan Taylor, and Robert
    Tibshirani. \href{http://arxiv.org/abs/1410.8260}{"Selecting the number of principal components: Estimation of the true rank of a noisy matrix."}
    arXiv preprint arXiv:1410.8260 (2014).
}
\author{
David Gerard
}
