% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_ests.R
\name{sig_soft}
\alias{sig_soft}
\title{Variance estimators from Choi et al (2014).}
\usage{
sig_soft(Y, c_val = 2/3, k = 10, lambda_grid = NULL,
  print_update = FALSE)
}
\arguments{
\item{Y}{The data matrix.}

\item{c_val}{The ad-hoc adjustment to the degrees of freedom. Choi et
al (2014) found that 2/3 worked well in simultations.}

\item{k}{A positive integer. The fold for the soft-impute cross validation. 
Default is 10.}

\item{lambda_grid}{A vector of positive numerics. The values of lambda to 
compute. The default is 20 values from the minimum to the maximum singular 
value of \code{Y}.}

\item{print_update}{A logical. Should we print to the screen the status of
the cross-validation-ish procedure at each iteration (TRUE) or not (FALSE)?}
}
\value{
\code{sig2_est} A positive numeric. The estimate of the
    variance.

    \code{sse} A positive numeric. The sum of squared errors for
    estimated Y.

    \code{dfLambda} A positive integer. The estimated df.
}
\description{
Returns a degrees of freedom corrected residual mean-squared error
of a soft-thresholding estimator where the tuning parameter is
chosen by a cross-validation-ish procedure.
}
\details{
See Choi et al (2014) for details. This seems to be the best
estimation procedure so far, but also takes the longest.

You can try out different values of \code{c_val} with the outputs
of \code{sse} and \code{dfLambda}
}
\references{
Choi, Yunjin, Jonathan Taylor, and Robert
    Tibshirani. \href{http://arxiv.org/abs/1410.8260}{"Selecting the number of principal components: Estimation of the true rank of a noisy matrix."}
    arXiv preprint arXiv:1410.8260 (2014).
}
\seealso{
\code{\link{soft_cv}} for the cross-validation-ish procedure.
}
\author{
David Gerard
}
