% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_coord.R
\name{soft_coord}
\alias{soft_coord}
\title{Runs an iterative coordinate descent algorithm to minimize the SURE for
mode-specific soft thresholding estimators.}
\usage{
soft_coord(c_obj, lambda_init, c_init, itermax = 1000, tol = 10^-4,
  print_iter = TRUE, tau2 = 1, use_sure = TRUE)
}
\arguments{
\item{c_obj}{The output from \code{\link{get_c}}.}

\item{lambda_init}{A vector of numerics of length \eqn{n}. The initial
starting values for the thresholding parameters.}

\item{c_init}{A positive numeric. The starting value of the scaling
parameter.}

\item{itermax}{A positive integer. The maximum number of Newton steps to
iterate through.}

\item{tol}{A positive numeric. The stopping criterion.}

\item{print_iter}{A logical. Should we print the updates of the Newton Step?}

\item{tau2}{A positive numeric. The variance. Assumed known and defaults to
1.}

\item{use_sure}{A logical. Which stopping criterion should we use? The mean
absolute difference in the parameters (\code{FALSE}) or the absolute value
of the deviation of the ratio of adjacent SURE values from 1
(\code{TRUE}).}
}
\value{
\code{c} A numeric. The final value of the scaling parameter.

  \code{lambda} A vector of numerics. The final values of the thresholding
  parameters.

  \code{est} An array of numerics. The final mean estimate.
}
\description{
Runs an iterative coordinate descent algorithm to minimize the SURE for
mode-specific soft thresholding estimators.
}
\references{
Gerard, D., & Hoff, P. (2015).
  \href{http://arxiv.org/abs/1505.02114}{Adaptive Higher-order Spectral
  Estimators}. \emph{arXiv preprint} arXiv:1505.02114.
}
\seealso{
\code{\link{update_c}}, \code{\link{update_lambda}},
  \code{\link{get_c}}.
}
\author{
David Gerard.
}
