% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_est.R
\name{stein}
\alias{stein}
\title{Stein's estimator.}
\usage{
stein(X, tau2 = 1)
}
\arguments{
\item{X}{A vector, matrix, or array of numerics. The data.}

\item{tau2}{A positive numeric. The variance. Assumed known and defaults to
1.}
}
\description{
Stein's estimator.
}
\author{
David Gerard.
}
\references{
James, W., & Stein, C. (1961, June).
  \href{http://projecteuclid.org/euclid.bsmsp/1200512173}{Estimation with
  quadratic loss}. \emph{In Proceedings of the fourth Berkeley symposium on
  mathematical statistics and probability} (Vol. 1, No. 1961, pp. 361-379).
}

