% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_sure.R
\name{sure}
\alias{sure}
\title{Wrapper for \code{\link{get_c}} and \code{\link{sure_given_c}}.}
\usage{
sure(X, func, dfunc, lambda, tau2 = 1)
}
\arguments{
\item{X}{An array of numerics. The data.}

\item{func}{A list of length \code{length(dim(X))} of shrinkage functions.}

\item{dfunc}{A list of length \code{length(dim(X))} of corresponding
derivative functions.}

\item{lambda}{A list of parameter values for shinking along each mode.}

\item{tau2}{A positive numeric. The variance. Assumed known and defaults to
1.}
}
\description{
Wrapper for \code{\link{get_c}} and \code{\link{sure_given_c}}.
}
\author{
David Gerard.
}
\references{
Gerard, D., & Hoff, P. (2015).
  \href{http://arxiv.org/abs/1505.02114}{Adaptive Higher-order Spectral
  Estimators}. \emph{arXiv preprint} arXiv:1505.02114.
}
\seealso{
\code{\link{get_c}}, \code{\link{sure_given_c}}.
}

