% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_coord.R
\name{sure_given_c_one_mode}
\alias{sure_given_c_one_mode}
\title{Wrapper for \code{\link{sure_given_c}} when only updating one mode.}
\usage{
sure_given_c_one_mode(lambda, c_obj, lambda_current, c_current, current_mode,
  tau2)
}
\arguments{
\item{lambda}{A numeric scalar. The current value of the lambda you are updating.}

\item{c_obj}{The output from \code{\link{get_c}}.}

\item{lambda_current}{A vector of numerics of length \eqn{n}. The current
values for the thresholding parameters.}

\item{c_current}{A positive numeric. The current value of the scaling
parameter.}

\item{current_mode}{The current mode for which lambda is a member.}

\item{tau2}{A positive numeric. The variance. Assumed known and defaults to
1.}
}
\description{
Wrapper for \code{\link{sure_given_c}} when only updating one mode.
}
\author{
David Gerard
}

