% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_est.R
\name{sure_matrix_complex}
\alias{sure_matrix_complex}
\title{SURE for spectral shrinkage functions for complex matrices.}
\usage{
sure_matrix_complex(d, lambda, p, tau2 = 1)
}
\arguments{
\item{d}{A vector of numerics. The singular values.}

\item{lambda}{A numeric. The shrinkage parameter.}

\item{p}{A vector of positive integers. The dimension of the data array.}

\item{tau2}{A positive numeric. The variance. Assumed known and defaults to
1.}
}
\description{
SURE for spectral shrinkage functions for complex matrices.
}
\author{
David Gerard
}
\references{
Candes, E., Sing-Long, C. A., & Trzasko, J. D. (2013).
  \href{http://ieeexplore.ieee.org/xpls/abs_all.jsp?arnumber=6545395&tag=1}{Unbiased
   risk estimates for singular value thresholding and spectral estimators}.
  \emph{Signal Processing, IEEE Transactions on}, 61(19), 4643-4657.
}

