% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_specific_sure.R
\name{sure_rank}
\alias{sure_rank}
\title{Iterates through all multilinear ranks less than \code{max_nrank} and
returns all SUREs.}
\usage{
sure_rank(X, max_nrank = dim(X), tau2 = 1)
}
\arguments{
\item{X}{An array of numerics. The data.}

\item{max_nrank}{A vector of positive integers. The maximum rank to inspect for each mode.}

\item{tau2}{The variance, assumed known. Defaults to 1.}
}
\value{
\code{min_rank} The multilinear rank that minimizes the SURE.

\code{min_sure} The minimum SURE value.

\code{sure_vec} A vector of all the SURE values of all the multilinear ranks that were inspected.

\code{all_ranks} A matrix of all of the multilinear ranks that were inspected.

\code{est} The final mean estimate.
}
\description{
Iterate through all ranks less than max_nrank and choose rank with smallest
sure.
}
\author{
David Gerard.
}
\references{
Gerard, D., & Hoff, P. (2015).
\href{http://arxiv.org/abs/1505.02114}{Adaptive Higher-order Spectral Estimators}.
\emph{arXiv preprint} arXiv:1505.02114.
}

