% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_sure.R
\name{sure_S}
\alias{sure_S}
\title{Wrapper for \code{\link{get_basics}} and \code{\link{sure_given_basics}}.}
\usage{
sure_S(X, func, dfunc, params, tau2 = 1)
}
\arguments{
\item{X}{An array of numerics. The tensor data.}

\item{func}{A shrinkage function.}

\item{dfunc}{The derivative function of \code{func}.}

\item{params}{The parameters to be fed into \code{func} and \code{dfunc}.}

\item{tau2}{A positive numeric. The variance. Assumed known and defaults to
1.}
}
\value{
\code{sure_val} The sure value of the higher-order spectral
  estimator.

  \code{est} The mean estimate.
}
\description{
Wrapper for \code{\link{get_basics}} and \code{\link{sure_given_basics}}.
}
\author{
David Gerard.
}
\references{
Gerard, D., & Hoff, P. (2015).
  \href{http://arxiv.org/abs/1505.02114}{Adaptive Higher-order Spectral
  Estimators}. \emph{arXiv preprint} arXiv:1505.02114.
}
\seealso{
\code{\link{get_basics}}, \code{\link{sure_given_basics}}.
}

