% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_ests.R
\name{tensor_var_est}
\alias{tensor_var_est}
\title{Non-parametric variance estimation for tensor datasets.}
\usage{
tensor_var_est(Y, sig_method = c("soft", "mp", "naive"), r = NULL)
}
\arguments{
\item{Y}{An array of numerics. The data tensor. Assumed to have a mean that
is low multilinear rank.}

\item{sig_method}{A string. The variance estimation procedure to use on each
matricization of the data tensor. Either "soft" for calling
\code{\link{sig_soft}}, "mp" for calling \code{\link{sig_mp}}, or "naive"
for calling \code{\link{sig_naive}}. See those functions for descriptions 
of these procedures.}

\item{r}{A vector of positive integers. The known multilinear rank of the
mean tensor. Only needed if \code{sig_method == "naive"}.}
}
\value{
\code{sig2_est} A positive numeric. The final variance estimate.
  
  \code{sig2_vec} A vector of positive numerics. The estimates of the
  variance when using each matricization of the data tensor. If these numbers
  are wildly different, then you should be skeptical of this variance
  estimation approach.
}
\description{
Implements one of three variance estimation procedures for matrix data sets
on each matricization of a data tensor. These estimates are then averaged to
come up with a final variance estimate.
}
\details{
If a data tensor has low multilinear rank mean, then this function will use
matrix-specific variance estimation procedues along each matricization of the
data tensor to estimate the variance. These estimates are then averaged to
provide the final estimate of the variance.
}
\author{
David Gerard
}
\references{
Choi, Yunjin, Jonathan Taylor, and Robert Tibshirani.
  \href{http://arxiv.org/abs/1410.8260}{"Selecting the number of principal
  components: Estimation of the true rank of a noisy matrix."} arXiv preprint
  arXiv:1410.8260 (2014).
  
  Gavish, M., & Donoho, D. L. (2014).
  \href{http://ieeexplore.ieee.org/xpls/abs_all.jsp?arnumber=6846297&tag=1}{The
   optimal hard threshold for singular values is 4/sqrt(3)}. Information
  Theory, IEEE Transactions on, 60(8), 5040-5053.
}
\seealso{
\code{\link{sig_soft}} for the soft-thresholding procedure,
  \code{\link{sig_mp}} for a procedure based on a specific asymptotic
  framework, and \code{\link{sig_naive}} for a basic estimator when the
  multilinear rank is known.
}

