% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cichocki_meet.R
\name{trunc_hosvd}
\alias{trunc_hosvd}
\title{Matrix-specific ways to estimate the rank for each mode and then return the
truncated HOSVD.}
\usage{
trunc_hosvd(Y, rank_vec = NULL, method = c("pa", "bcv", "mdl"),
  return_est = TRUE)
}
\arguments{
\item{Y}{A numeric array.}

\item{rank_vec}{The multilinear rank of the underlying mean. If \code{NULL} then
the multilinear rank will be estimated.}

\item{method}{The way to estimate the multilinear rank if \code{rank_vec = NULL}.
 The methods allowed are parallel analysis (\code{"pa"}) from
 \code{\link[sva]{num.sv}}, bicrossvalidation (\code{"bcv"})
from \code{\link[cate]{est.factor.num}},
 and minimum description length (\code{"mdl"}) from \code{\link{mdl_eigen}}.}

\item{return_est}{A logical. Should we return the final estimate from the
truncated HOSVD (\code{TRUE}) or note (\code{FALSE})?}
}
\description{
Matrix-specific ways to estimate the rank for each mode and then return the
truncated HOSVD.
}
\author{
David Gerard
}
