% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_coord.R
\name{update_c}
\alias{update_c}
\title{Update scale parameter in scale and soft-thresholding HOSE's.}
\usage{
update_c(c_obj, lambda_current, c_current, tau2, epsilon = 10^-4)
}
\arguments{
\item{c_obj}{The output from \code{\link{get_c}}.}

\item{lambda_current}{A vector of numerics of length \eqn{n}. The initial
starting values for the thresholding parameters.}

\item{c_current}{A positive numeric. The starting value of the scaling
parameter.}

\item{tau2}{A positive numeric. The variance. Assumed known and defaults to
1.}

\item{epsilon}{A positive numeric. The min distance of the Newton update.}
}
\value{
\code{c_new} A positive numeric. The update of the scaling parameter.
}
\description{
Update scale parameter in scale and soft-thresholding HOSE's.
}
\references{
Gerard, D., & Hoff, P. (2015).
  \href{http://arxiv.org/abs/1505.02114}{Adaptive Higher-order Spectral
  Estimators}. \emph{arXiv preprint} arXiv:1505.02114.
}
\seealso{
\code{\link{soft_coord}}.
}
\author{
David Gerard.
}
