% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_coord.R
\name{update_lambda}
\alias{update_lambda}
\title{Update of thresholding parameter for scale and soft-thresholding HOSEs.}
\usage{
update_lambda(c_obj, lambda_current, c_current, k, tau2, epsilon = 10^-4)
}
\arguments{
\item{c_obj}{The output from \code{\link{get_c}}.}

\item{lambda_current}{A vector of numerics of length \eqn{n}. The current
values for the thresholding parameters.}

\item{c_current}{A positive numeric. The current value of the scaling
parameter.}

\item{k}{A positive integer. The mode to update.}

\item{tau2}{A positive numeric. The variance. Assumed known and defaults to
1.}

\item{epsilon}{A positive numeric. The min distance of the Newton update.}
}
\value{
\code{lambda_new} A numeric. The update of the threshholding
  parameter.
}
\description{
Update of thresholding parameter for scale and soft-thresholding HOSEs.
}
\author{
David Gerard.
}
\references{
Gerard, D., & Hoff, P. (2015).
  \href{http://arxiv.org/abs/1505.02114}{Adaptive Higher-order Spectral
  Estimators}. \emph{arXiv preprint} arXiv:1505.02114.
}
\seealso{
\code{\link{soft_coord}}.
}

