% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_coord.R
\name{update_lambda_brent}
\alias{update_lambda_brent}
\title{Apply Brent's method to update the kth mode when soft-thresholding.}
\usage{
update_lambda_brent(c_obj, lambda_current, c_current, k, tau2)
}
\arguments{
\item{c_obj}{The output from \code{\link{get_c}}.}

\item{lambda_current}{A vector of numerics of length \eqn{n}. The current
values for the thresholding parameters.}

\item{c_current}{A positive numeric. The current value of the scaling
parameter.}

\item{k}{A positive integer. The mode to update.}

\item{tau2}{A positive numeric. The variance. Assumed known and defaults to
1.}
}
\description{
Apply Brent's method to update the kth mode when soft-thresholding.
}
\author{
David Gerard
}

