% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage_functions.R
\name{f_S_lasso}
\alias{f_S_lasso}
\title{Soft thresholding a core array.}
\usage{
f_S_lasso(S, lambda)
}
\arguments{
\item{S}{An array of numerics.}

\item{lambda}{A numeric. The thresholding parameter.}
}
\description{
Soft thresholding a core array.
}
\seealso{
\code{\link{df_S_lasso}}.
}
\author{
David Gerard.
}
